requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-toolbelt==0.0.1
djangorestframework==3.3.1
fake-factory==0.5.3
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
